lint: preserve runtime package file-count baseline
